// platform/frontend/src/client/types.gen.ts
export type HistoryMessage = {
    /**
     * Role
     *
     * The role of the message sender
     */
    role: 'user' | 'assistant';
    /**
     * Content
     *
     * The content of the message
     */
    content: string;
};